refactor(drive-integration): rename inspectGoogleDoc → fetchDocumentScope and prune backward-compat dead code [INTEG-4233]
