+++ apps/drive-integration/src/utils/fetchDocumentScope.ts
